car-game-new2/backend/requirements.txt
flask
mysql-connector-python

